+++ src/main/java/uk/gov/ida/matchingserviceadapter/mappers/MatchingDatasetToMatchingDatasetDtoMapper.java
@@ -4,6 +4,7 @@ import com.google.common.base.Function;
 import com.google.common.collect.Collections2;
 import com.google.common.collect.Lists;
 import org.joda.time.LocalDate;
+import uk.gov.ida.matchingserviceadapter.Comparators;
 import uk.gov.ida.matchingserviceadapter.rest.matchingservice.GenderDto;
 import uk.gov.ida.matchingserviceadapter.rest.matchingservice.SimpleMdsValueDto;
 import uk.gov.ida.matchingserviceadapter.rest.matchingservice.TransliterableMdsValueDto;
@@ -18,7 +19,6 @@ import uk.gov.ida.saml.core.domain.SimpleMdsValue;
 import uk.gov.ida.saml.core.domain.TransliterableMdsValue;
 
 import javax.annotation.Nullable;
-import java.util.Comparator;
 import java.util.List;
 import java.util.Optional;
 import java.util.stream.Collectors;
@@ -26,7 +26,7 @@ import java.util.stream.Collectors;
 public class MatchingDatasetToMatchingDatasetDtoMapper {
 
     public VerifyMatchingDatasetDto mapToVerifyMatchingDatasetDto(MatchingDataset matchingDataset) {
-        Optional<TransliterableMdsValue> firstNameValue = matchingDataset.getFirstNames().stream().min(comparatorByVerifiedThenCurrent());
+        Optional<TransliterableMdsValue> firstNameValue = matchingDataset.getFirstNames().stream().min(Comparators.comparatorByVerifiedThenCurrent());
         Optional<SimpleMdsValue<String>> middleNameValue = matchingDataset.getMiddleNames().stream().findFirst();
         Optional<SimpleMdsValue<LocalDate>> birthDateValue = matchingDataset.getDateOfBirths().stream().findFirst();
 
@@ -114,10 +114,5 @@ public class MatchingDatasetToMatchingDatasetDtoMapper {
                         input.getTo(),
                         input.isVerified()))
                 .collect(Collectors.toList());
-    }
-
-    public static Comparator<TransliterableMdsValue> comparatorByVerifiedThenCurrent() {
-        return Comparator.comparing(TransliterableMdsValue::isVerified, Comparator.reverseOrder())
-                .thenComparing(TransliterableMdsValue::getTo, Comparator.nullsFirst(null));
     }
 }
