+++ src/test/java/uk/gov/ida/matchingserviceadapter/mappers/MatchingDatasetToMatchingDatasetDtoMapperTest.java
@@ -82,28 +82,6 @@ public class MatchingDatasetToMatchingDatasetDtoMapperTest {
     }
 
     @Test
-    public void comparatorPrioritisesVerifiedThenCurrent() {
-        List<TransliterableMdsValue> firstNames = new ArrayList<>();
-        firstNames.add(buildFirstName("historical unverified: expected seventh", DateTime.now(), false));
-        firstNames.add(buildFirstName("current unverified: expected fifth", null, false));
-        firstNames.add(buildFirstName("historical verified: expected third", DateTime.now(), true));
-        firstNames.add(buildFirstName("current verified: expected first", null, true));
-        firstNames.add(buildFirstName("historical unverified: expected eighth", DateTime.now(), false));
-        firstNames.add(buildFirstName("current unverified: expected sixth", null, false));
-        firstNames.add(buildFirstName("historical verified: expected fourth", DateTime.now(), true));
-        firstNames.add(buildFirstName("current verified: expected second", null, true));
-        firstNames.sort(MatchingDatasetToMatchingDatasetDtoMapper.comparatorByVerifiedThenCurrent());
-        assertEquals("current verified: expected first", firstNames.get(0).getValue());
-        assertEquals("current verified: expected second", firstNames.get(1).getValue());
-        assertEquals("historical verified: expected third", firstNames.get(2).getValue());
-        assertEquals("historical verified: expected fourth", firstNames.get(3).getValue());
-        assertEquals("current unverified: expected fifth", firstNames.get(4).getValue());
-        assertEquals("current unverified: expected sixth", firstNames.get(5).getValue());
-        assertEquals("historical unverified: expected seventh", firstNames.get(6).getValue());
-        assertEquals("historical unverified: expected eighth", firstNames.get(7).getValue());
-    }
-
-    @Test
     public void shouldMapToUniversalMatchingDatasetDto() {
         LocalDate dob = new LocalDate(1970, 1, 2);
         LocalDate oldDob = new LocalDate(1970, 2, 1);
@@ -153,10 +131,5 @@ public class MatchingDatasetToMatchingDatasetDtoMapperTest {
         assertThat(matchingDatasetDto.getFirstName()).contains(new TransliterableMdsValueDto("Joe", null, null, null, true));
         assertThat(matchingDatasetDto.getSurnames()).containsOnly(new TransliterableMdsValueDto("Bloggs", null, null, null, true));
         assertThat(matchingDatasetDto.getDateOfBirth()).contains(new SimpleMdsValueDto<>(dob, null, null, true));
-    }
-
-    private TransliterableMdsValue buildFirstName(String name, DateTime to, boolean verified) {
-        SimpleMdsValue<String> simpleValue = new SimpleMdsValue<>(name, null, to, verified);
-        return new TransliterableMdsValue(simpleValue);
     }
 }
